style: move codeGen and selector to builders folder
